+++ app/src/main/java/org/sopt/cgv/feature/time/component/TimeScreenMovieSelectionSection.kt
@@ -1,4 +1,4 @@
-package org.sopt.cgv.feature.time
+package org.sopt.cgv.feature.time.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
